Portal: Fix onOutClick after child re-render.
* The previous implementation triggered `onOutClick` after a click in
the `Portal` content if it gets re-rendered during that click. It
assumed that if the clicked element is not found in the root element
via `root.contains(e.target)`, it's outside. But if after re-render the
clicked element gets removed from the DOM, so it cannot be found in the
root element. Instead we capture and flag the click event before it
bubbles up
to the `document` to be handled by `handleOutClick`.

* It's recommended to use `ref` callbacks instead of `findDOMNode`.
https://github.com/yannickcr/eslint-plugin-react/blob/master/docs/rules/
no-find-dom-node.md

* Extra `handleOutClick` on the prototype is not needed, it is only
used if `canUseDOM` and in that case gets overwritten in the
`constructor`.